webapp/requirements.txt
# Thai OCR Web Application Dependencies

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# Deep Learning
torch>=1.9.0
transformers>=4.20.0

# Tokenizer
sentencepiece>=0.1.96

# Image Processing
Pillow>=8.0.0

# HTTP Requests (for URL image loading)
requests>=2.25.0
